Added EffortLeft field to the Task PopOver when adding logs. Automatically updates when duration is changed and is set to default when duration is invalid, or set to 0 if the duration goes past the effort left.

+++ src/main/java/seng302/group2/workspace/project/story/tasks/Log.java
@@ -36,7 +36,7 @@ public class Log extends SaharaItem implements Serializable {
      * @param duration the duration the person worked for
      * @param startTime time the logger started working on the task
      */
-    public Log(Task task, String description, Person logger, Float duration, LocalDate startTime) {
+    public Log(Task task, String description, Person logger, double duration, LocalDate startTime) {
         super("Untitled Log");
         this.task = task;
         this.logger = logger;
